refactor: reduce the Participation facade to a dispatch table
Last of the four. The two roster reads join Sessions now that Attendance and
SessionNotes exist to serve them, and completing a session becomes
CompleteSession.execute/2 -- the one write that spans two entities under one
transaction, which is the rule provider/offboard_staff_member.ex states for
living at the context root rather than inside a sub-domain module.

participation.ex ends at 173 lines: 39 defdelegate, zero def, zero defp. It
drops `use KlassHero.Shared.Tracing` and every alias but the five it dispatches
to, matching provider.ex, whose moduledoc this one now echoes.

Dropping the facade's tracing does not lose the context: `context_span` reports
context.name from the *second* module segment, not the last, so
Participation.Attendance still reports "Participation" -- fixed in #1424 for
exactly this shape, after Provider's delegate facade turned out to emit no
Provider spans at all. Both live Honeycomb triggers filter on context.name, so
neither is affected. The free-form span *names* do change for all 16 spanned
operations, from Participation.<op> to Participation.<Module>.<op>.

program_name/1 folds back to private, as its doc said it would: it was public
only for the roster read that has now moved in beside it.

The context ends up at 2219 lines against 1899 before -- a real ~17% expansion,
paid in five moduledocs, five alias blocks, and documentation that now sits with
the implementation. What it buys is a per-file defp:def ratio in line with every
other context: the facade was carrying 123 private functions against 46 public
ones, where Messaging carries 7 and Provider none.
